experimental/transformers-to-ctranslate/requirements.txt
ctranslate2
huggingface_hub
transformers
python-dotenv
